// src/api/schema.d.ts
export interface components {
    schemas: {
        LoginUser: {
            email: string;
            /** Format: password */
            password: string;
        };
        NewUser: {
            username: string;
            email: string;
            /** Format: password */
            password: string;
        };
        User: {
            email: string;
            token: string;
            username: string;
            bio: string;
            image: string;
        };
        UpdateUser: {
            email?: string;
            password?: string;
            username?: string;
            bio?: string;
            image?: string;
        };
        Profile: {
            username: string;
            bio: string;
            image: string;
            following: boolean;
        };
        Article: {
            slug: string;
            title: string;
            description: string;
            body: string;
            tagList: string[];
            /** Format: date-time */
            createdAt: string;
            /** Format: date-time */
            updatedAt: string;
            favorited: boolean;
            favoritesCount: number;
            author: components["schemas"]["Profile"];
        };
        NewArticle: {
            title: string;
            description: string;
            body: string;
            tagList?: string[];
        };
        UpdateArticle: {
            title?: string;
            description?: string;
            body?: string;
        };
        Comment: {
            id: number;
            /** Format: date-time */
            createdAt: string;
            /** Format: date-time */
            updatedAt: string;
            body: string;
            author: components["schemas"]["Profile"];
        };
        NewComment: {
            body: string;
        };
        GenericErrorModel: {
            errors: {
                body: string[];
            };
        };
    };
// ...
}